+++ tests/extras/test_torch.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
-'''Test operator identities.'''
-import numpy as np
+'''Test torch extra.'''
 from numpy.random import default_rng
 from numpy.testing import assert_allclose
 import pytest
 
 from eminus import Atoms, config
-config.use_torch = False
+config.use_torch = True
+assert config.use_torch
 
 # Create an Atoms object to build mock wave functions
 atoms = Atoms('Ne', [0, 0, 0], ecut=1).build()
@@ -19,31 +19,9 @@ W_tests = {
     'full_spin': rng.standard_normal((atoms.Nspin, len(atoms.G2), atoms.Nstate)),
     'active_spin': rng.standard_normal((atoms.Nspin, len(atoms.G2c), atoms.Nstate))
 }
-dr = rng.standard_normal(3)
 
 
-@pytest.mark.parametrize('type', ['full', 'full_spin'])
-def test_LinvL(type):
-    out = atoms.Linv(atoms.L(W_tests[type]))
-    test = np.copy(W_tests[type])
-    if test.ndim == 3:
-        test[:, 0, :] = 0
-    else:
-        test[0, :] = 0
-    assert_allclose(out, test)
-
-
-@pytest.mark.parametrize('type', ['full', 'full_spin'])
-def test_LLinv(type):
-    out = atoms.L(atoms.Linv(W_tests[type]))
-    test = np.copy(W_tests[type])
-    if test.ndim == 3:
-        test[:, 0, :] = 0
-    else:
-        test[0, :] = 0
-    assert_allclose(out, test)
-
-
+@pytest.mark.extras
 @pytest.mark.parametrize('type', ['full', 'full_single', 'full_spin'])
 def test_IJ(type):
     out = atoms.I(atoms.J(W_tests[type]))
@@ -51,6 +29,7 @@ def test_IJ(type):
     assert_allclose(out, test)
 
 
+@pytest.mark.extras
 @pytest.mark.parametrize('type', ['full', 'active', 'full_single', 'active_single', 'full_spin',
                                   'active_spin'])
 def test_JI(type):
@@ -62,6 +41,7 @@ def test_JI(type):
     assert_allclose(out, test)
 
 
+@pytest.mark.extras
 @pytest.mark.parametrize('type', ['active', 'active_single', 'active_spin'])
 def test_IdagJdag(type):
     out = atoms.Idag(atoms.Jdag(W_tests[type]))
@@ -69,6 +49,7 @@ def test_IdagJdag(type):
     assert_allclose(out, test)
 
 
+@pytest.mark.extras
 @pytest.mark.parametrize('type', ['full', 'full_single', 'full_spin'])
 def test_JdagIdag(type):
     out = atoms.Jdag(atoms.Idag(W_tests[type], True))
@@ -76,9 +57,58 @@ def test_JdagIdag(type):
     assert_allclose(out, test)
 
 
+@pytest.mark.extras
+@pytest.mark.parametrize('type', ['full', 'full_single', 'full_spin'])
+def test_IJ_gpu(type):
+    try:
+        config.use_gpu = True
+        assert config.use_gpu
+    except AssertionError:
+        pytest.skip('GPU not available, skip tests')
+    out = atoms.I(atoms.J(W_tests[type]))
+    test = W_tests[type]
+    assert_allclose(out, test)
+
+
+@pytest.mark.extras
+@pytest.mark.parametrize('type', ['full', 'active', 'full_single', 'active_single', 'full_spin',
+                                  'active_spin'])
+def test_JI_gpu(type):
+    try:
+        config.use_gpu = True
+        assert config.use_gpu
+    except AssertionError:
+        pytest.skip('GPU not available, skip tests')
+    if 'active' in type:
+        out = atoms.J(atoms.I(W_tests[type]), False)
+    else:
+        out = atoms.J(atoms.I(W_tests[type]))
+    test = W_tests[type]
+    assert_allclose(out, test)
+
+
+@pytest.mark.extras
 @pytest.mark.parametrize('type', ['active', 'active_single', 'active_spin'])
-def test_TT(type):
-    out = atoms.T(atoms.T(W_tests[type], dr), -dr)
+def test_IdagJdag_gpu(type):
+    try:
+        config.use_gpu = True
+        assert config.use_gpu
+    except AssertionError:
+        pytest.skip('GPU not available, skip tests')
+    out = atoms.Idag(atoms.Jdag(W_tests[type]))
+    test = W_tests[type]
+    assert_allclose(out, test)
+
+
+@pytest.mark.extras
+@pytest.mark.parametrize('type', ['full', 'full_single', 'full_spin'])
+def test_JdagIdag_gpu(type):
+    try:
+        config.use_gpu = True
+        assert config.use_gpu
+    except AssertionError:
+        pytest.skip('GPU not available, skip tests')
+    out = atoms.Jdag(atoms.Idag(W_tests[type], True))
     test = W_tests[type]
     assert_allclose(out, test)